Melhorias
Animations em desuso apagados, novo script para Hitbox, testes novos com classe Enemy etc.

+++ Cronicas da Bin. Game Project/Assets/Hitbox.cs
@@ -40,7 +40,7 @@ public class Hitbox : MonoBehaviour
     {
         currentHealth -= damage;
 
-        hitRenderer.SetState(currentHealth);        
+        hitRenderer.SetState((int) currentHealth);        
 
         if (currentHealth <= 0)
         {
